trimming some code (slide mechanism)
