+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/ActiveIntakeSubsystem.java
@@ -1,0 +1,62 @@
+package org.firstinspires.ftc.teamcode.subsystems;
+
+import com.arcrobotics.ftclib.command.SubsystemBase;
+import com.qualcomm.robotcore.eventloop.opmode.OpMode;
+import com.qualcomm.robotcore.hardware.CRServo;
+import com.qualcomm.robotcore.hardware.DcMotor;
+import com.qualcomm.robotcore.hardware.DcMotorEx;
+import com.qualcomm.robotcore.hardware.HardwareMap;
+import com.arcrobotics.ftclib.hardware.ServoEx;
+import com.arcrobotics.ftclib.hardware.SimpleServo;
+import com.qualcomm.robotcore.util.Range;
+
+import org.firstinspires.ftc.teamcode.Constants;
+import org.firstinspires.ftc.teamcode.Constants.IntakeConstants;
+
+import java.lang.Math;
+
+public class IntakeSubsystem extends SubsystemBase {
+    private HardwareMap hardwareMap;
+
+    private final ServoEx wristServo;
+    private final CRServo activeIntakeServo;
+
+    private double speedMultiplier = 1.0;
+    private final OpMode opMode;
+
+    public IntakeSubsystem(HardwareMap hardwareMap,OpMode opMode){
+        this.opMode= opMode;//a little fix so that the subsystem itself can add things into telemetry
+
+        // wrist servo creation
+        wristServo = new SimpleServo(hardwareMap,IntakeConstants.INTAKE_WRIST_SERVO_NAME, IntakeConstants.INTAKE_WRIST_SERVO_MIN_ANGLE,IntakeConstants.INTAKE_WRIST_SERVO_MAX_ANGLE);
+        activeIntakeServo = hardwareMap.get(CRServo.class, IntakeConstants.ACTIVE_INTAKE_SERVO_NAME);
+    }
+    
+    public void ActiveIntakeServoIn(){
+        activeIntakeServo.setPower(1.0);
+    }
+
+    public void ActiveIntakeServoOut(){
+        activeIntakeServo.setPower(-1.0);
+    }
+
+    public void stopActiveIntakeServo(){
+        activeIntakeServo.setPower(0.0);
+    }
+
+    /*
+    moves the wrist servo into position to retract
+    */
+    public void servoUpPosition(){
+        wristServo.turnToAngle(IntakeConstants.INTAKE_WRIST_SERVO_UP_POSITION);
+    }
+
+    /*
+    moves the wrist servo into position to pick up samples
+     */
+    public void servoDownPosition(){
+        wristServo.turnToAngle(IntakeConstants.INTAKE_WRIST_SERVO_DOWN_POSITION);
+    }
+
+
+}
